Fixed type_for_source logic
